feat: added Shoe Controller and edit resource

+++ src/main/java/com/acme/myshoes/platform/shoes/resource/ShoeResource.java
@@ -10,5 +10,6 @@ import lombok.*;
 public class ShoeResource {
     Long id;
     String name;
-    Float Size;
+    int Size;
+
 }
